fix vlimits code (#38)
1.  ProcessInFifo.c, VlimitInLookup.c, authindi.c, authpgsql.c,
    iauth.c, sql_getpw.c, vadduser.c, vchkpass.c, vlimit.c, vmoduser.c,
    vsetuserquota.c: initialize struct vlimits
2. limits.c: fix data types to fix stack smashing
3. vlimits.h: changed data types to long
4. vlimit.c: added -T option to toggle bit field
5. vlimit.c: fixed data types
6. inquerytest: change data type to long
7. vadduser.c: fix .domain_limits path

+++ indimail-x/vlimits.h
@@ -1,5 +1,8 @@
 /*
  * $Log: vlimits.h,v $
+ * Revision 1.4  2024-05-27 08:05:16+05:30  Cprogrammer
+ * changed data type of variables to long
+ *
  * Revision 1.3  2024-05-17 16:25:48+05:30  mbhangui
  * fix discarded-qualifier compiler warnings
  *
@@ -30,11 +33,11 @@ struct vlimits {
       /* max service limits */
       long      domain_expiry;
       long      passwd_expiry;
-      int       maxpopaccounts;
-      int       maxaliases;
-      int       maxforwards;
-      int       maxautoresponders;
-      int       maxmailinglists;
+      long      maxpopaccounts;
+      long      maxaliases;
+      long      maxforwards;
+      long      maxautoresponders;
+      long      maxmailinglists;
 
       /* quota & message count limits */
       umdir_t   diskquota;			/*- domain quota */
